Cast it back

+++ src/live_restore/live_restore_fs.c
@@ -1704,7 +1704,7 @@ __wt_os_live_restore_fs(
     /* Configure the read size. */
     WT_ERR(__wt_config_gets(session, cfg, "live_restore.read_size", &cval));
     lr_fs->read_size = (uint64_t)cval.val;
-    if (!__wt_ispo2(lr_fs->read_size))
+    if (!__wt_ispo2((uint32_t)lr_fs->read_size))
         WT_ERR_MSG(session, EINVAL, "the live restore read size must be a power of two");
 
     /* Update the callers pointer. */
